Convert to Cargo workspace and extract git-tidy-core
Restructure the single crate into a workspace to enable sharing
classification logic and the git abstraction layer between
git-worktree-tidy and future companion tools (git-branch-tidy).

- Create git-tidy-core library crate with shared modules:
  git.rs, types.rs, error.rs, classification.rs, dirty.rs, landed.rs
- Consolidate MockGitBuilder/MockGit/TestRepo into testutil.rs,
  gated behind a "testutil" feature flag
- Move worktree-specific modules (cli, discovery, output, clean)
  to crates/git-worktree-tidy/
- Update all imports: core modules use crate::, binary uses
  git_tidy_core:: for shared types
- All 94 existing tests pass unchanged
