+++ Cargo.toml
@@ -1,21 +1,3 @@
-[package]
-name = "git-worktree-tidy"
-version = "0.1.0"
-edition = "2024"
-rust-version = "1.85.0"
-description = "Scan, classify, and interactively remove stale Git worktrees"
-
-[dependencies]
-clap = { version = "4", features = ["derive"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-dialoguer = "0.11"
-console = "0.15"
-thiserror = "2"
-strsim = "0.11"
-toml = "0.8"
-
-[dev-dependencies]
-assert_cmd = "2"
-predicates = "3"
-tempfile = "3"
+[workspace]
+members = ["crates/*"]
+resolver = "3"
