+++ crates/git-tidy-core/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "git-tidy-core"
+version = "0.1.0"
+edition = "2024"
+description = "Shared classification logic and git abstraction for git-worktree-tidy and git-branch-tidy"
+
+[features]
+testutil = ["dep:tempfile"]
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+strsim = "0.11"
+thiserror = "2"
+toml = "0.8"
+
+[dev-dependencies]
+tempfile = "3"
+
+[dependencies.tempfile]
+version = "3"
+optional = true
