+++ crates/git-tidy-core/src/config.rs
