+++ crates/git-tidy-core/src/error.rs
