+++ crates/git-tidy-core/src/landed.rs
@@ -301,7 +301,7 @@ mod tests {
     use std::path::PathBuf;
 
     use super::*;
-    use crate::git::tests::MockGitBuilder;
+    use crate::testutil::MockGitBuilder;
 
     fn repo() -> PathBuf {
         PathBuf::from("/repo")
