+++ crates/git-tidy-core/src/lib.rs
@@ -1,11 +1,10 @@
 pub mod classification;
-pub mod clean;
-pub mod cli;
 pub mod config;
 pub mod dirty;
-pub mod discovery;
 pub mod error;
 pub mod git;
 pub mod landed;
-pub mod output;
 pub mod types;
+
+#[cfg(any(test, feature = "testutil"))]
+pub mod testutil;
