+++ crates/git-tidy-core/src/types.rs
