+++ crates/git-worktree-tidy/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod clean;
+pub mod cli;
+pub mod discovery;
+pub mod output;
+
+// Re-export core crate for convenience
+pub use git_tidy_core;
