+++ crates/git-worktree-tidy/tests/integration_discovery.rs
